Move token tranforms config to transforms/ directory
