+++ tokens/transformation/android/colorName.js
