+++ tokens/transformation/android/formatFontStyle.js
