+++ tokens/transformation/android/formatResourcesSorted.js
