+++ tokens/transformation/common/camelCaseHelper.js
