+++ tokens/transformation/common/colorToHex8.js
