+++ tokens/transformation/common/colorToRgbaString.js
