+++ tokens/transformation/common/copyFileOrFolder.js
