+++ tokens/transformation/ios/fontStyleTemplate.js
