+++ tokens/transformation/web/createPropertyFormatter.js
