+++ tokens/transformation/web/formattedVariables.js
